Attempt setup to handle different native video sources; doesn't work
The source picker looks to find the smallest amount of scale, but the problem is that some sources would technically be scale=1 (because that's the scale required to not crop the region of interest) but they would be scaled up. This is because "scale" does not refer to how much the video itself will be scaled, but rather what percentage of the viewport width we will render the element at. A correctly functioning source picker would look at the video's native dimensions AND the scaled size.